Refactor Code Editor
