+++ src/frontoffice-frontend/src/constants/routes.js
@@ -7,4 +7,6 @@ export const SCRIPT_PAGE = {route: "/script/:id"}
 
 export const CHALLENGES_PAGE = {route: "/challenge/"}
 
+export const CHALLENGE_PAGE = {route: "/challenge/:id"}
+
 export const TOURNAMENTS_PAGE = {route: "/tournament/"}
