+++ src/frontoffice-frontend/src/views/script-page/index.js
@@ -15,38 +15,9 @@ import {listenUsersLeft} from "../../requests/websocket/script/consumer/userLeft
 import {listenUsersJoined} from "../../requests/websocket/script/consumer/userJoinedScript";
 import {joinScript} from "../../requests/websocket/script/producer/joinScript";
 import {leaveScript} from "../../requests/websocket/script/producer/leaveScript";
-
-const languages = [
-    {
-        id: 'ccd651ad-012d-4c01-8287-43e117f75de6',
-        key: "javascript",
-        icon: <i className="devicon-nodejs-plain" style={{fontSize: 22}}></i>,
-        sample: `/* Simple Hello World in Node.js */
-console.log("Hello World");
-`
-    },
-    {
-        id: 'f7f620d1-cf38-431e-9640-9a6a05dea978',
-        key: "python",
-        icon: <i className="devicon-python-plain" style={{fontSize: 22}}></i>,
-        sample: `# Simple Hello World in Python
-print("Hello World")
-`
-    },
-    {
-        id: '35ecd470-d2eb-4e57-b9ae-970a8686b7bd',
-        key: "text/x-java",
-        icon: <i className="devicon-java-plain" style={{fontSize: 22}}></i>,
-        sample: `// Simple Hello World in Java
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello World");
-    }
-}
-`
-    },
-];
-
+import {sendChange} from "../../requests/websocket/script/producer/sendScriptChange";
+import {listenScriptChanges} from "../../requests/websocket/script/consumer/scriptChange";
+import {LANGUAJES} from "../../constants/languages";
 
 const ScriptPage = () => {
     const { id } = useParams()
@@ -56,7 +27,7 @@ const ScriptPage = () => {
     const [formTab, setFormTab] = useState("login")
     const [scriptUsers, setScriptUsers] = useState([])
 
-    const [language, setLanguage] = useState(languages[0]);
+    const [language, setLanguage] = useState(LANGUAJES[0]);
     const [title, setTitle] = useState('');
     const [shareKey, setShareKey] = useState('');
 
@@ -96,13 +67,31 @@ const ScriptPage = () => {
 
         findScriptById(id).then((response) => {
             setTitle(response.data.name);
-            setLanguage(languages.find(language => language.id === response.data.languageId));
+            setLanguage(LANGUAJES.find(language => language.id === response.data.languageId));
             setShareKey(response.data.shareKey);
         }).catch(e =>
             errorNotification("Error retrieving script", e.response.data.message || "Try again later",
                 "topRight")
         );
     }, []);
+
+
+    // WHEN CODE CHANGES
+    useEffect(() => {
+        editorRef.current.on('change', (instance, changes) => {
+            setCode(editorRef.current.getValue());
+            if(changes.origin !== "setValue" && changes.origin !== undefined) {
+                sendChange(stompClient, id, localStorage.getItem("username"), changes, instance.getValue());
+            }
+        });
+    }, [setCode, stompClient]);
+
+    useEffect(() => {
+        listenScriptChanges(stompClient, id, (change, codeChanged) => {
+            editorRef.current.replaceRange(change.text, change.from, change.to, 'setValue');
+            setCode(codeChanged);
+        });
+    }, [stompClient]);
 
     if(!isAuthenticated) {
         return(
@@ -116,10 +105,10 @@ const ScriptPage = () => {
                   scriptUsers={scriptUsers} setLanguage={setLanguage} />
             <Layout>
                 <Sider width={"fit-content"}>
-                    <CodeEditorOptions id={id} language={language.key} setLanguage={setLanguage} languages={languages} />
+                    <CodeEditorOptions id={id} language={language.key} setLanguage={setLanguage} languages={LANGUAJES} />
                 </Sider>
                 <Content style={{ padding: '24px 24px', height: '100vh' }}>
-                    <CodeEditor scriptId={id} language={language.key} setCode={setCode} editorRef={editorRef} stompClient={stompClient} />
+                    <CodeEditor language={language.key} editorRef={editorRef} />
                 </Content>
                 <Sider width={"fit-content"} style={{padding: '24px 24px', backgroundColor: '#F5F5F5', height: '100vh'}}>
                     <CodeRunBox
